211005 1053 이혜미
위치정보 필터링, 페이징, 검색 기능 구현

+++ gooddog/src/main/java/com/gooddog/domain/WeightVO.java
@@ -10,7 +10,7 @@ public class WeightVO {
 	// 펫 몸무게 번호
 	private int weight_no;
 	// 펫 몸무게
-	private int weight_content;
+	private float weight_content;
 	// 몸무게 측정일시
 	private String weight_date;
 	// 펫 번호
